reparación controladores y refactoring nombres clases
