+++ src/main/java/cl/praxis/bootcapp/controllers/RoleController.java
@@ -2,9 +2,7 @@ package cl.praxis.bootcapp.controllers;
 
 
 import cl.praxis.bootcapp.entities.Role;
-import cl.praxis.bootcapp.services.services.IBaseServiceCRUD;
-import cl.praxis.bootcapp.services.services.imp.RoleServiceImp;
-import org.springframework.beans.factory.annotation.Autowired;
+import cl.praxis.bootcapp.services.imp.RoleServiceImp;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -12,12 +10,13 @@ import org.springframework.web.bind.annotation.GetMapping;
 import java.util.List;
 
 @Controller
-public class RolesController {
+public class RoleController {
 
-    @Autowired
-    private IBaseServiceCRUD roleRepository;
-    @Autowired
-    private RoleServiceImp roleServiceImp;
+   private RoleServiceImp roleServiceImp;
+
+    public RoleController(RoleServiceImp roleServiceImp) {
+        this.roleServiceImp = roleServiceImp;
+    }
 
     @GetMapping()
     public String getAllRoles(Model model) {
