+++ src/main/java/cl/praxis/bootcapp/services/IBaseServiceCRUD.java
@@ -1,4 +1,4 @@
-package cl.praxis.bootcapp.services.services;
+package cl.praxis.bootcapp.services;
 
 import java.util.List;
