+++ src/main/java/cl/praxis/bootcapp/services/IRoleCRUD.java
@@ -1,4 +1,4 @@
-package cl.praxis.bootcapp.services.services;
+package cl.praxis.bootcapp.services;
 
 
 import cl.praxis.bootcapp.entities.Role;
