+++ src/main/java/cl/praxis/bootcapp/services/IUserService.java
@@ -1,0 +1,4 @@
+package cl.praxis.bootcapp.services;
+
+public interface IUserService {
+}
