+++ src/main/java/cl/praxis/bootcapp/services/imp/UserServiceImpl.java
@@ -1,9 +1,9 @@
-package cl.praxis.bootcapp.services.services.imp;
+package cl.praxis.bootcapp.services.imp;
 
 import cl.praxis.bootcapp.entities.User;
 import cl.praxis.bootcapp.repositories.IUserRepository;
-import cl.praxis.bootcapp.services.services.IBaseServiceCRUD;
-import cl.praxis.bootcapp.services.services.IUserService;
+import cl.praxis.bootcapp.services.IBaseServiceCRUD;
+import cl.praxis.bootcapp.services.IUserService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
