--- src/main/java/cl/praxis/bootcapp/services/services/IUserService.java
@@ -1,4 +1,0 @@
-package cl.praxis.bootcapp.services.services;
-
-public interface IUserService {
-}
